% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorInput.R
\name{scaffoldTemplates}
\alias{scaffoldTemplates}
\title{Scaffold handlebars template files.}
\usage{
scaffoldTemplates(path = ".")
}
\arguments{
\item{path}{Where should template files be written?}
}
\description{
Creates local copies of the handlebars templates used to render the widget in
the client.  This is useful for customizing the appearance of the widget.
See the `template` argument in \code{\link{vectorInput}}.
}
