% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorInput.R
\name{updateVectorInput}
\alias{updateVectorInput}
\title{Update the values in a vectorInput widget}
\usage{
updateVectorInput(session, inputId, values)
}
\arguments{
\item{session}{Shiny server sesssion object.}

\item{inputId}{Target inputId.}

\item{values}{A new set of values.}
}
\description{
Update the values in a vectorInput widget
}
