% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorInput.R
\name{vectorInput}
\alias{vectorInput}
\title{vectorInput}
\usage{
vectorInput(inputId, type = "text", nameLabel = "name",
  valueLabel = "value", values = NULL, template = if (is.na(nameLabel)) { 
     file.path(templatePath, "unnamed-vector.html") } else {    
  file.path(templatePath, "named-vector.html") })
}
\arguments{
\item{inputId}{The Shiny inputId}

\item{type}{An <input> element type.  See Shiny and jsonlight documentation
for conversion rules.}

\item{nameLabel}{A label to use for the column of value names.  Set to NA to
create an unnamed vector.}

\item{valueLabel}{A label to use for the column of values.}

\item{values}{An initial set of values.  Names are ignored if nameLabel = NA.}

\item{template}{A file path to a handlebars template to use to render the
widget.  Use \code{\link{scaffoldTemplates}} to create a local copy of the
default templates.}
}
\value{
A shiny tagList.
}
\description{
A shiny input widget that returns a (possibly named) vector of values.
}
